+++ dev/tools/Message.cs
@@ -6,22 +6,11 @@ using ct = dein.tools.Colorify.Type;
 
 namespace dein.tools
 {
-    public static class Message
+    public class Message
     {
-        private static Config _c { get; set; }
-        private static PersonalConfiguration _cp { get; set; }
-
-        static Message()
-        {
-            _c = Program.config;
-            _cp = Program.config.personal;
-        }
-
         public static void Critical(string msg = null){
-            _cp.mnu.sel = "m";
             Error(msg, !String.IsNullOrEmpty(msg));
         }
-
 
         public static void Error(string msg = null, bool replace = false, bool exit = false) {
             Colorify.Default();
@@ -82,10 +71,8 @@ namespace dein.tools
                 $" Press [Any] key to continue...".txtWarning();
                 Console.ReadKey();
 
-                if (!exit)
+                if (exit)
                 {
-                    Menu.Start();
-                } else {
                     Program.Exit();
                 }
             }
@@ -98,7 +85,7 @@ namespace dein.tools
             Colorify.Default();
             Console.Clear();
             
-            bool opt_cnf = false;        
+            bool opt_cnf = false;
             try
             {
                 $"=".bgWarning(ct.Repeat);
@@ -109,10 +96,14 @@ namespace dein.tools
                 $" {msg}".txtDefault(ct.Shell);
 
                 $"".fmNewLine();
+                $"{"[EMPTY] Cancel", 82}".txtDanger(ct.WriteLine);
+
+                $"".fmNewLine();
                 $"=".bgWarning(ct.Repeat);
                 $"".fmNewLine();
 
                 $" [Y] Yes or [N] No: ".txtWarning();
+
                 string opt = Console.ReadLine();
                 switch (opt?.ToLower())
                 {
@@ -121,6 +112,9 @@ namespace dein.tools
                         break;
                     case "n":
                         opt_cnf = false;
+                        break;
+                    case "":
+                        //Menu.Route();
                         break;
                     default:
                         Message.Error();