+++ dev/tools/Validation.cs
@@ -18,9 +18,7 @@ namespace dein.tools
             }
             catch (Exception Ex)
             {
-                Message.Critical(
-                    msg: Ex.Message
-                );
+                Exceptions.General(Ex.Message);
             }
         }
 
@@ -38,9 +36,7 @@ namespace dein.tools
             }
             catch (Exception Ex)
             {
-                Message.Critical(
-                    msg: Ex.Message
-                );
+                Exceptions.General(Ex.Message);
             }
         }
 
@@ -61,9 +57,7 @@ namespace dein.tools
             }
             catch (Exception Ex)
             {
-                Message.Critical(
-                    msg: Ex.Message
-                );
+                Exceptions.General(Ex.Message);
             }
         }
     }