+++ dev/view/Android.cs
@@ -436,11 +436,8 @@ namespace HardHat {
                         StringBuilder msg = new StringBuilder();
                         msg.Append($"There is a new Android Build Tools version installed.");
                         msg.Append(Environment.NewLine);
-                        msg.Append($" Do you want change ANDROID_BT_VERSION from {currentVersion} to {lastVersion}?");
-                        bool change = Message.Confirmation(msg.ToString());
-                        if (change){
-                            Variables.Value("ab", lastVersion);
-                        }
+                        msg.Append($" Please verify your Environment Variables and change ANDROID_BT_VERSION from {currentVersion} to {lastVersion}.");
+                        Message.Alert(msg.ToString());
                     }
                 }
             }