+++ dev/view/Gulp.cs
@@ -77,8 +77,6 @@ namespace HardHat {
         }
         
         public static void Select() {
-            Gulp.Check();
-            
             Colorify.Default();
             Console.Clear();
             try
@@ -127,7 +125,7 @@ namespace HardHat {
                 Section.SelectedProject();
                 Section.CurrentConfiguration(_cp.mnu.g_val, _cp.mnu.g_cnf);
 
-                Protocols.Start();
+                Selector.Start(Selector.Protocol, "1");
 
                 string opt_ptc = Console.ReadLine();
                 opt_ptc = opt_ptc?.ToLower();
@@ -548,8 +546,7 @@ namespace HardHat {
                 Section.HorizontalRule();
                 Section.Pause();
 
-                Menu.Status();
-                Select();
+                Menu.Start();
             }
             catch (Exception Ex){
                 Exceptions.General(Ex.Message);