+++ CHANGELOG.md
@@ -9,6 +9,17 @@ Fixed       for any bug fixes.
 Security    to invite users to upgrade in case of vulnerabilities.
 -->
 
+## [4.1.1] - 2017-11-09
+
+### Fixed
+
+* Gulp update redirection fix.
+* One instance running check from startup.
+
+### Changed
+
+* Build Tools update message now only works as alert.
+
 ## [4.1.0] - 2017-11-08
 
 ### Added